phase3/requirements.txt
# Phase 3 - Dépendances avec Tests
# Installation: pip install -r requirements.txt

# LLM Framework
anthropic>=0.40.0

# Data Validation
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0

# Kafka
confluent-kafka>=2.3.0

# RAG & Vector Store
chromadb>=0.4.22
sentence-transformers>=2.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-timeout>=2.1.0
